+++ src/Sulu/Component/Content/Document/Structure/ManagedStructure.php
@@ -52,7 +52,7 @@ class ManagedStructure extends Structure
     /**
      * {@inheritDoc}
      */
-    public function getProperty($name)
+    public function getProperty($name, $locale = null)
     {
         $this->init();
 
@@ -69,7 +69,10 @@ class ManagedStructure extends Structure
         $contentTypeName = $structureProperty->getType();
 
         if ($structureProperty->isLocalized()) {
-            $locale = $this->inspector->getLocale($this->document);
+            if ($locale === null) {
+                $locale = $this->inspector->getLocale($this->document);
+            }
+
             $property = $this->legacyPropertyFactory->createTranslatedProperty($structureProperty, $locale);
         } else {
             $property = $this->legacyPropertyFactory->createProperty($structureProperty);