+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-master
     * HOTFIX      #1568 [ContentBundle]  Fixed copy extension data
+    * HOTFIX      #1567 [Document]       Fixed localized property for url property
 
 * 1.0.8 (2015-08-31)
     * HOTFIX      #1539 [WebsiteBundle]  Fixed canonical tag for shadow pages